responsive FIT_ALL layout
If last double-click/tap action was FIT_ALL, then register an event listener on the resize action such that the layout is redrawn after a window resize.  Remove the event listener for all other div resizing actions so that in all cases but FIT_ALL a window resize doesn't change the widths of the columns/divs.