removed 'click' event listener after init
Modified 'click' event listener to immediately perform FIT_ALL action, and then deregister itself as an event listener.  I'm concerned about performance impacts, especially considering that implementation as it stood was re-adding the resizer handles on every click, overlaying them on top of one another.  I could imagine that after keeping a page up for a long time and interacting with it through many clicks, that could lead to issues.